refactor: rename job.mountain → job.mountain.basic
